+++ src/FlubuCore/Tasks/Solution/VSSolutionBrowsing/ProjectType.cs
@@ -1,0 +1,10 @@
+﻿namespace FlubuCore.Tasks.Solution.VSSolutionBrowsing
+{
+    public enum ProjectType
+    {
+        Unknown,
+        LibraryOrConsoleApp,
+        WebApp,
+        Service,
+    }
+}